backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
django-filter==22.1
Pillow==9.2.0
gunicorn==20.1.0
python-dotenv==0.21.0
whitenoise==6.2.0